(fixed) the logic on deleting a post

+++ social-media-app/backend/src/controllers/posts.controllers.js
@@ -1,4 +1,4 @@
-import { sendCreated, sendDeleteSuccess, sendNotFound, sendServerError } from "../helpers/helper.function.js";
+import { forbidden, notAuthorized, sendCreated, sendDeleteSuccess, sendNotFound, sendServerError } from "../helpers/helper.function.js";
 import { createPostService, deleteAPostService, getAllPostsService, getOnePostService, updateAPostService } from "../services/postService.js";
 
 
@@ -6,6 +6,7 @@ import { createPostService, deleteAPostService, getAllPostsService, getOnePostSe
 
 export const createPost=async (req,res)=>{
     try{
+         
          const newPost={
               user_id:req.user.user_id,
               content:req.body.content,
@@ -56,15 +57,33 @@ export const getOnePost =async (req,res)=>{
 export const updateAPost=async(req, res)=>{
     try {
         let post_id=req.params.post_id;
-        let {content}=req.body
-        const response= await updateAPostService(post_id,content);
-        console.log(response)
-        if(response.rowsAffected[0]==1){    
-             sendDeleteSuccess(res,"updated successfully")
+        let user_id=req.user.user_id;
+        let postDetails=await getOnePostService(post_id);
+
+        if(postDetails.length==0){
+            sendNotFound(res,"post does not exist")
         }
+        else{
+            if(postDetails[0].user_id==req.user.user_id){
+                let {content}=req.body
+            
+                const response= await updateAPostService(post_id,content);
+                console.log(response)
 
+                if(response.rowsAffected[0]>0){    
+                    sendSuccess(res,"updated successfully")
+                 }
+                 else{
+                    sendNotFound(res,"post not found ")
+                  
+                }
 
-        
+            }else{
+                forbidden(res,'forbidden')
+            }
+
+        }
+               
     } catch (error) {
         sendServerError(res,error)
     }
@@ -72,21 +91,40 @@ export const updateAPost=async(req, res)=>{
 
 export const deleteAPost=async(req,res)=>{
     try {
+        let user_id=req.user.user_id
+        console.log('user_id',user_id)
+
         let post_id=req.params.post_id;
+        //check whether the user is the owner of the post
 
-        let response =await deleteAPostService(post_id)
+        let postDetails=await getOnePostService(post_id);
 
-        // console.log(response);
-        if(response.rowsAffected[0]==1){    
-            sendDeleteSuccess(res,"deleted successfully")
+        console.log("This is a post details",postDetails)
+
+        if(postDetails.length==0){
+            sendNotFound(res,"post does not exist")
+        }else{
+            if(postDetails[0].user_id==req.user.user_id){
+                let response =await deleteAPostService(post_id)
+
+                // console.log(response);
+                if(response.rowsAffected>0){    
+                    sendDeleteSuccess(res,"deleted successfully")
+                }
+                else{
+                    sendNotFound(res,"post not found or it has already been deleted")
+                }
+
+            }
+            else{
+                forbidden(res,'forbidden')
+            }
+
         }
-        else{
-            sendNotFound(res,"post not found or it has already been deleted")
-        }
-        
+     
         
     } catch (error) {
-        sendServerError(res, error)   
+        sendServerError(res, error.message)   
     }
 }
